Add a tool to convert introspection data to HardwareDetails
This adds a tool called "make-introspection" that takes 2 arguments,
the Ironic inspector URI and then node UUID. It queries inspector
for the data, and if found, converts it to the BareMetalHost
HardwareDetails format.
